Refactor view data tests for heading
There were a lot of identical tests for heading content across implementing
classes, and as this value should always be identical in each implementation,
centralising these two cases to the new bast class test.
