+++ server/pages/appointments/attendanceOutcomePage.test.ts
@@ -301,14 +301,12 @@ describe('AttendanceOutcomePage', () => {
 
     it('returns view data for a session', () => {
       const session = sessionFactory.build()
-      const formattedDate = '10 June 2026'
       const form = appointmentOutcomeFormFactory.build({
         contactOutcome: contactOutcomeFactory.build({ code: contactOutcomes[1].code }),
       })
       const notesItems = { notes: 'Test notes', showIsSensitiveQuestion: false }
 
       jest.spyOn(paths.sessions, 'update')
-      jest.spyOn(DateTimeFormats, 'isoDateToUIDate').mockReturnValue(formattedDate)
       jest.spyOn(NotesUtils, 'questionItems').mockReturnValue(notesItems)
 
       const page = new AttendanceOutcomePage({
@@ -348,19 +346,14 @@ describe('AttendanceOutcomePage', () => {
         page: 'choose-supervisor',
       })
 
-      expect(result).toEqual({
-        backLink: pathWithQuery,
-        updatePath: pathWithQuery,
-        form: undefined,
-        heading: {
-          title: `${session.projectName}(${formattedDate})`,
-          caption: 'Bulk update',
-        },
-        ...notesItems,
-        items: expectedItems,
-      })
-
-      expect(DateTimeFormats.isoDateToUIDate).toHaveBeenCalledWith(session.date)
+      expect(result).toEqual(
+        expect.objectContaining({
+          backLink: pathWithQuery,
+          updatePath: pathWithQuery,
+          ...notesItems,
+          items: expectedItems,
+        }),
+      )
     })
 
     it('passes undefined appointment to questionItems when appointmentOrSession is a session', () => {
