+++ server/pages/appointments/logHoursPage.test.ts
@@ -1,17 +1,13 @@
 import { AppointmentDto } from '../../@types/shared'
-import Offender from '../../models/offender'
 import paths from '../../paths'
 import appointmentFactory from '../../testutils/factories/appointmentFactory'
 import sessionFactory from '../../testutils/factories/sessionFactory'
 import attendanceDataFactory from '../../testutils/factories/attendanceDataFactory'
 import LogHoursPage, { LogHoursQuery } from './logHoursPage'
 import * as Utils from '../../utils/utils'
-import DateTimeFormats from '../../utils/dateTimeUtils'
 import { AppointmentOutcomeForm } from '../../@types/user-defined'
 import appointmentOutcomeFormFactory from '../../testutils/factories/appointmentOutcomeFormFactory'
 import { contactOutcomeFactory } from '../../testutils/factories/contactOutcomeFactory'
-
-jest.mock('../../models/offender')
 
 describe('LogHoursPage', () => {
   let page: LogHoursPage
@@ -232,7 +228,6 @@ describe('LogHoursPage', () => {
     let appointment: AppointmentDto
     let form: AppointmentOutcomeForm
     const updatePath = '/update'
-    const offenderMock: jest.Mock = Offender as unknown as jest.Mock<Offender>
 
     beforeEach(() => {
       page = new LogHoursPage()
@@ -375,25 +370,6 @@ describe('LogHoursPage', () => {
       })
     })
 
-    it('should return an object containing heading details', () => {
-      const offender = {
-        name: 'Sam Smith',
-        crn: 'CRN123',
-        isLimited: false,
-      }
-
-      offenderMock.mockImplementation(() => {
-        return offender
-      })
-
-      const result = page.viewData(appointment, form)
-
-      expect(result.heading).toEqual({
-        title: offender.name,
-        caption: offender.crn,
-      })
-    })
-
     it('should return an object containing a back link to the attendance outcome page', async () => {
       jest.spyOn(paths.appointments, 'update')
       const result = page.viewData(appointment, form)
@@ -439,10 +415,6 @@ describe('LogHoursPage', () => {
 
       expect(result.backLink).toBe(pathWithQuery)
       expect(result.updatePath).toBe(pathWithQuery)
-      expect(result.heading).toEqual({
-        title: `${session.projectName}(${DateTimeFormats.isoDateToUIDate(session.date)})`,
-        caption: 'Bulk update',
-      })
     })
   })
